Implement Save Data
This commit implements the ability to save data. The top-performing bot will save their transition matrix to a file, which can be used as a starting-off point later. Additionally, this can be used to create a bot which can play against human players, which should represent the current best bot in play.